Add files via upload
Todo App UI Mobile using HTML CSS.
About this Project
In this project you will be developing the application in a way that same website can work on multiple devices having different screen sizes in pixels. To achieve this you will be using Media Queries, one of the finest concept in CSS to build responsive applications.